code review changes

+++ scripts/local-install.sh
@@ -1,17 +1,26 @@
 #!/usr/bin/env bash
 set -eo pipefail
 
-## Example:
+## usage: local-install.sh <package-to-install-dir> <consuming-app-dir>'
 ##
-## local-install.sh sdk/js-client-sdk-common sdk/js-client-sdk && local-install.sh sdk/js-client-sdk test-apps/my-test-react-app
+## examples:
 ##
+##  Install js-client-sdk-common changes into js-client-sdk:
+##   ./scripts/local-install.sh . ../js-client-sdk
+##
+##  Install js-client-sdk changes into a test app
+##   ./scripts/local-install.sh ../js-client-sdk ../test-apps/my-react-app
+##
+##  Install js-client-sdk-common changes into a test app
+##   ./scripts/local-install.sh . ../js-client-sdk && ./scripts/local-install.sh ../js-client-sdk ../test-apps/my-react-app
 
-## (requires jq (`brew install jq`))
+if ! command -v jq &> /dev/null; then
+  echo "jq must be installed before using this script"
+  exit 1
+fi
 
 if [[ "$1" = "" || "$2" = "" ]]; then
-  echo 'USAGE: ./local-install.sh <package-to-install-path> <consuming-app-path>'
-  ehco ''
-  echo ''
+  echo 'usage: local-install.sh <package-to-install-dir> <consuming-app-dir>'
   exit 1
 fi
 
@@ -30,15 +39,20 @@ if [ ! -d "$2" ]; then
   exit 1
 fi
 
-rm -rf ~/.yarn/berry/cache/\@eppo-*
+if [ ! -f "$2/package.json" ]; then
+  echo "$2 is not an npm package"
+  exit 1
+fi
 
 ### Create local package via "npm pack"
 pushd "$1" > /dev/null
-rm -rf package pack.out pack.err && npm pack 2> pack.err > pack.out
-COMPRESSED_PACKAGE="$(tail -1 pack.out)"
+mkdir -p /tmp/packages
+PACKAGE_DIR=/tmp/packages/"$(basename "$(pwd)")"
+rm -rf "${PACKAGE_DIR}" /tmp/pack.out && npm pack > /tmp/pack.out
+COMPRESSED_PACKAGE="$(tail -1 /tmp/pack.out)"
 tar -xzf "${COMPRESSED_PACKAGE}"
 rm "${COMPRESSED_PACKAGE}"
-PACKAGE_DIR="$(pwd)/package"
+mv package "${PACKAGE_DIR}"
 PACKAGE_NAME=$(cat ./package.json | jq -r '.name')
 popd > /dev/null
